+++ pyautolabnova/autolab.py
@@ -1,6 +1,0 @@
-# In autolab.py within pyautolabnova
-from directfunctions import DirectFunctions
-from generalfunctions import GeneralFunctions
-
-class Autolab(DirectFunctions, GeneralFunctions):
-    pass