+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0] - 2024-07-28
+
+### Added
+- Basic functionality for connecting to and controlling Autolab device
+- Autolab control method for setting mode, current range, and bandwidth
+- Cell on/off functionality
+
+### Changed
+- Moved from pre-alpha to alpha stage
+- Improved error handling and connection process
+
+### Removed
+- Temporarily removed non-functioning methods (apply_current, record_signals)
+
 ## [0.0.1a4] - 2024-07-22
 
 ### Added
@@ -16,5 +30,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Unable to disconnect from device
 
+## [0.0.1a3] - 2024-07-18
+
+### Added
+- Entry file
+
+### Changed
+- Improved entry file structure
+
+### Fixed
+- Static paths for support files
+
 ### Removed
 - dll_path_finder.py as well as other unused files
+
+## [0.0.1a1] - 2024-07-18
+
+### Added
+- Entry file
+
+### Changed
+- N/A
+
+### Fixed
+- N/A
+
+### Removed
+- N/A